Feat (experiment)!: Try Yazi's approach of creating a tmp dir to extract to
This change is completely experimental, and will be immediately reverted
in the next commit. I just want to get this into version control. The
approach of creating a temporary directory to extract to isn't that
great, since there is a lot of clean up to do and files to manage. The
next commit will completely revert this change and use 7z to test the
extraction before it actually extracts the archive, which removes the
need for the temporary directory, making things far less complex and
easier to maintain.

If you want to try this commit, below are the breaking changes.

Breaking changes:
- Extraction behaviour is now either "rename" or "overwrite", because
  the plugin is no longer using 7zip to decide what to do with the
  extracted files, and is using Yazi's fs.unique_name function to create
  a unique name for the extracted files if "rename" is wanted.
  "Overwrite" mode will just copy the files over without creating a
  unique name for it, which would overwrite the existing files.