Feat (revert experiment)!: Test the archive before actually extracting it
This commit is a breaking change with respect to the previous commit,
but if we leave the previous commit out, there is no configuration
change needed.

Breaking changes (with respect to the previous commit):
- The plugin now supports all 7z extraction behaviours instead of only
  support "overwrite" and "rename" like in the previous commit.

Testing the archive before actually extracting it allows the plugin to
not need to create temporary directories to extract the archive to and then
moving the files to the right place, and then removing them
while still not polluting the directory with the failed empty files.
It is much simpler and less complex to manage,
which is great for maintainability.