Feat: Warn the user about invalid configuration options
This change is to make configuring the plugin slightly easier since the
plugin has a lot of configuration options. Some of which are very
long as well, so having a message that tells the user about the invalid
configuration options that they have given should help catch typos and
legacy configuration options, which should be changed or removed.