Feat: Add support for extracting archives recursively
Changes:
- No change is required on the user side, as the default option is the
  to have this feature enabled, as this feature was available for
  archives containing only a single file.
- Added a configuration option to disable extracting archives
  recursively.
- Added documentation about the new extract_archives_recursively
  configuration option.

The way the plugin recursively extracts archive files is by listing out
all files in the archive, then selecting the files that end with an
archive file extension. The plugin then checks if the file is actually
an archive with the file command, then proceeds to extract the archive.

This means the plugin will not work on archive files inside the initial
archive file that do not have the correct file extension. This
limitation is so that the plugin does not have to call the file command
on every single file that has been extracted from the archive to check
if it is an archive file or not, which would be very slow and very
unnecessary, as most files in an archive are not an archive themselves.

There is minor performance hit with this change, since the plugin
now has to parse the entire output of the 7z l command to get the
list of all the items in the archive, so that it knows the path to the
archive files in the directory containing the extracted items from the
archive.

This plugin might feel a bit more sluggish as well, since the directory
change only occurs after all archive files have been extracted, not just
the initial archive file.