Feat: Make use of the new ya.confirm API
This is only supported on the latest version of Yazi from the main
branch. Stable versions of Yazi currently do not have this API, and will
fall back to using ya.input.

The ya.input fallback will be removed once the ya.confirm API is available
in a stable release, which will bump the plugin's minimum Yazi version.