Fix: Return only the escaped replacement string
The previous version of the escape replacement string function
just naively returned the result of the substitution, which included a
count of the number of replacements. This broke the handle_shell command
as it meant that the number of times to substitute the shell variable
was set to 0 if the % sign wasn't found.

Should have typed the function from the beginning, seems like I forgot
to do that. That would have avoided this issue entirely.