Fix: Fix is_archive_mime_type function returning an extractor class
It should be returning a boolean, not a class. I'm not sure how come
this didn't break with the default 7zip extractor, since both classes
inherit from the same base Extractor class, so the stored data inside
each class is similar and only consists of Lua primitives. It's weird
that it breaks for the Tar extractor but not for the 7zip extractor.

I forgot to update the functions making use of the
ARCHIVE_MIME_TYPE_TO_EXTRACTOR_MAP as that variable was just renamed
from the previous variable name, ARCHIVE_MIME_TYPES.

Also removed "application/octet-stream" as a mime type, as Arch Linux
has backported the patch from the main branch of file.

Bug fix source:
https://gitlab.archlinux.org/archlinux/packaging/packages/file/-/commit/8004cbb712a714a4ce834c9d5dee6e1c8dbf8ff1

Fixes #10.