Feat: Add confirmation password prompt
This confirmation password prompt will only apply when creating an
encrypted archive. The plugin will prompt you for the password twice,
and if both of the passwords do not match, it will show an error and
prompt you for passwords again. Cancelling either of the prompts will
cancel the entire process.